==> 3_text-classification.ipynb <==
import pandas as pd
books=pd.read_csv("books_cleaned.csv")

# Remember that books has too many categories (479)
books["categories"].value_counts().reset_index().sort_values("count",ascending=False)


# Find categories > 50
books["categories"].value_counts().reset_index().query("count>50")

# Let's rename some categories {<originalCategory>:<newlyMappedCategory>}
# Done by observing description of each category
category_mapping={'Fiction': "Fiction",
                  'Juvenile Fiction': "Children's Fiction",
                  'Biography & Autobiography': "Nonfiction",
                  'History': "Nonfiction",
                  'Literary Criticism': "Nonfiction",
                  'Philosophy': "Nonfiction",
                  'Religion': "Nonfiction",
                  'Comics & Graphic Novels': "Fiction",
                  'Drama': "Fiction",
                  'Juvenile Nonfiction': "Children's Nonfiction",
                  'Science': "Nonfiction",
                  'Poetry': "Fiction"}

# Create new column to do the mapping
books["simple_categories"]=books["categories"].map(category_mapping)


books

# How many books have simple_category not Nan?
books[~(books["simple_categories"].isna())]

# Install package torch
# import torch...?

# Now let's use a pre-trained model without modifying anything in it, maybe just add a few lines to the prompt
# So let's use a zero-shot classification model from hugging face
# Model: facebook/bart-large-mnli

# Before that, let's change the model download path to D:
import os
os.environ["TRANSFORMERS_CACHE"] = "D:/huggingface_cache"

# Now go to: Use this model > transformers > copy code

# Use a pipeline as a high-level helper
from transformers import pipeline
pipe = pipeline("zero-shot-classification", model="facebook/bart-large-mnli")

fiction_categories=["Fiction","Nonfiction"]

# Let's look at the description of the first known fiction example that we have
# Since the rows will have the original indices, let's reset them to start in order (from 0 obv)
sequence=books.loc[books["simple_categories"]=="Fiction","description"].reset_index(drop=True)[0]

# Into pipe, pass the sequence and categories to classify it into
pipe(sequence,fiction_categories)

import numpy as np

# Store the index of the higher score in a variable called max_index
max_index=np.argmax(pipe(sequence,fiction_categories)["scores"])
max_index # 0, as index 0 contains the higher score: 'scores': [0.8438268899917603, 0.15617306530475616]

# Store the label with the higher score in a variable called max_label
max_label=pipe(sequence,fiction_categories)["labels"][max_index]
max_label # Fiction

# Let's wrap it up in a function
def generate_predictions(sequence,categories):
    predictions=pipe(sequence,categories)
    max_index=np.argmax(predictions["scores"])
    max_label=predictions["labels"][max_index]
    return max_label

# Now the model will do zero shot classification over a description and tell if the label is fiction or non fiction

# Model validation
from tqdm import tqdm # Python library used to show progress bars
actual_categories=[]
predicted_categories=[]

# For fiction
for i in tqdm(range(300)):
    sequence=books.loc[books["simple_categories"]=="Fiction","description"].reset_index(drop=True)[i]
    predicted_categories+=[generate_predictions(sequence,fiction_categories)]
    actual_categories+=["Fiction"]

# For non fiction
for i in tqdm(range(300)):
    sequence=books.loc[books["simple_categories"]=="Nonfiction","description"].reset_index(drop=True)[i]
    predicted_categories+=[generate_predictions(sequence,fiction_categories)]
    actual_categories+=["Nonfiction"]

predictions_df=pd.DataFrame({"actual_categories":actual_categories,"predicted_categories":predicted_categories})
predictions_df

# Add a column called correct_prediction that has value 1 if actual and predicted are same, else 0
predictions_df["correct_prediction"]=(
    np.where(predictions_df["actual_categories"]==predictions_df["predicted_categories"], 1, 0))

predictions_df["correct_prediction"].sum()/len(predictions_df) # If all predictions are correct, then all values would be 1, then sum()==len() => sum\len==1

About 78% is correct

isbn=[]
predicted_categories=[]

# Find the isbn and description of rows with missing categories
missing_categories=books.loc[books["simple_categories"].isna(), ["isbn13", "description"]].reset_index(drop=True)
missing_categories

# Find the predicted categories for the actually missing categories
for i in tqdm(range(len(missing_categories))):
    sequence=missing_categories["description"][i]
    predicted_categories+=[generate_predictions(sequence, fiction_categories)]
    isbn+=[missing_categories["isbn13"][i]]

# Create a new dataframe that contains the isbn and predicted category of rows with actually missing categories
missing_predicted_df=pd.DataFrame({"isbn13":isbn,"predicted_categories":predicted_categories})
missing_predicted_df

# Merge these dataframes (left join missing_predicted_df onto books
books=pd.merge(books,missing_predicted_df,on="isbn13",how="left")

# Now books df contains a new column predicted_categories (from missing_predicted_df)
# When the simple_categories column value is missing, should use the value from predicted_categories, else keep the same value as in simple_categories
books["simple_categories"]=np.where(books["simple_categories"].isna(),books["predicted_categories"],books["simple_categories"])

# Now predicted_categories column can be dropped
books=books.drop(columns=["predicted_categories"])
books

- Now the books categories column is fixed (almost)
- On statistical observation, fictional category is about 54 % (I think)
- So we can try to break it down maybe

books[books["categories"].str.lower().isin([
    "romance",
    "science fiction"
    "scifi",
    "Fantasy",
    "horror",
    "mystery",
    "thriller",
    "comedy",
    "crime",
    "historical"
    ])]

- We could go on and try to classify based on these, but right now let's stop here

books

# Save the dataframe
(books
    .to_csv("books_with_categories.csv",index=False)
 )
