==> 4_sentiment-analysis.ipynb <==
import pandas as pd
books=pd.read_csv("books_with_categories.csv")

# We will use an open-source model from HF for text\emotional classification

# Before that, let's change the model download path to D:
import os
os.environ["TRANSFORMERS_CACHE"] = "D:/huggingface_cache"

# Model: j-hartmann/emotion-english-distilroberta-base

from transformers import pipeline
classifier = pipeline("text-classification",
                      model="j-hartmann/emotion-english-distilroberta-base",
                      # return_all_scores=True,
                      top_k=None)
classifier("I love this!")

# Let's see what the first book's description is
books["description"][0]

# Now let's run sentiment analysis over this description
classifier(books["description"][0])

- Maximum score seems to be fear, but on reading it doesn't seem like fear
- It feels more like a mixture of emotions
- So instead let's split the description into individual sentences and classify based on that

# Split description sentence wise and classify
sentences=books["description"][0].split(".")
predictions=classifier(sentences)
predictions # output shows sentence wise emotional score

# Can compare each sentence with the predicted emotion and check
for i in range(5):
    print("Sentence:\n",sentences[i])
    print("Emotion:\n",classifier(sentences[i]))
    print()

# Let's sort the emotions for the first prediction by label
sorted(predictions[0],key=lambda x: x["label"])

# For each description, we should find the maximum emotion score of each row
# Then we should return the maximum of those max values

import numpy as np
emotion_labels = ["anger", "disgust", "fear", "joy", "sadness", "surprise", "neutral"]
isbn = []
emotion_scores={label: [] for label in emotion_labels}

def calculate_max_emotion_scores(predictions):
    per_emotion_scores={label: [] for label in emotion_labels}
    for prediction in predictions:
        # Sort the emotions of the prediction by label
        sorted_predictions=sorted(prediction,key=lambda x: x["label"])

        for index, label in enumerate(emotion_labels):
            per_emotion_scores[label].append(sorted_predictions[index]["score"])
    return {label: np.max(scores) for label, scores in per_emotion_scores.items()}

from tqdm import tqdm

emotion_labels = ["anger", "disgust", "fear", "joy", "sadness", "surprise", "neutral"]
isbn = []
emotion_scores={label: [] for label in emotion_labels}

for i in tqdm(range(len(books))):
    isbn.append(books["isbn13"][i])
    sentences = books["description"][i].split(".")
    predictions = classifier(sentences)
    max_scores = calculate_max_emotion_scores(predictions)
    for label in emotion_labels:
        emotion_scores[label].append(max_scores [label])

emotions_df=pd.DataFrame(emotion_scores)
emotions_df["isbn13"]=isbn
emotions_df.head() # Now there is a separate column for each emotion

# Merge
books=pd.merge(books,emotions_df,on="isbn13")

# Save
books.to_csv("books_with_emotions.csv",index=False)
